backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
httpx==0.27.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
openai==1.51.0
